fix(registry): refactoring and fix validation
