+++ client/src/modules/Registry/components/index.tsx
@@ -1,0 +1,5 @@
+export * from './MentorStepRegistration';
+export * from './MentorStepAbout';
+export * from './MentorStepSuccess';
+export * from './MentorRegistrationGeneral';
+export * from './MentorRegistrationMentorship';
